# events/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vale night - Eventos",
        "description": "Micro serviços de eventos/festas",
        "contact": {
            "email": "arthurabreu00@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "local host server"
        }
    ],
    "paths": {
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get events informations",
                "description": "Returns events",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update event",
                "description": "Returns event",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the event",
                                        "type": "string",
                                        "example": "Nome do evento"
                                    },
                                    "description": {
                                        "description": "Description of the event",
                                        "type": "string",
                                        "example": "Descriçãp da festa"
                                    },
                                    "highlight": {
                                        "description": "Is highlight event",
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "age_group": {
                                        "description": "Age group of event",
                                        "type": "string",
                                        "example": "Jovens"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Update event",
                "description": "Returns event",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/events": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "description": "Returns event",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the event",
                                        "type": "string",
                                        "example": "Nome do evento"
                                    },
                                    "description": {
                                        "description": "Description of the event",
                                        "type": "string",
                                        "example": "Descriçãp da festa"
                                    },
                                    "highlight": {
                                        "description": "Is highlight event",
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "age_group": {
                                        "description": "Age group of event",
                                        "type": "string",
                                        "example": "Jovens"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "200": {
                "description": "successful operation"
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Event id",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        }
    }
}